removed files not used and organize the files in folders
