+++ rosa/src/components/Dedicating.js
